+++ src/impl/NodoABB.java
@@ -3,7 +3,7 @@ package impl;
 import api.ArbolBBTDA;
 
 public class NodoABB {
-    int Info;
+    double Info;
     ArbolBBTDA HijoIzq;
     ArbolBBTDA HijoDer;
